// src/env.d.ts
export declare const env: {
  AUTH_SECRET: string | undefined;
  AUTH_DISCORD_ID: string;
  AUTH_DISCORD_SECRET: string;
  DATABASE_URL: string;
  EMAIL_FROM: string;
  SMTP_HOST: string;
  SMTP_PORT: number;
  SMTP_USER: string;
  SMTP_PASSWORD: string;
  SMTP_SECURE: boolean;
  NODE_ENV: "development" | "test" | "production";
};